@@ -1,13 +1,14 @@
 module github.com/OliveTin/OliveTin
 
-go 1.23.4
+go 1.24
 
-toolchain go1.23.7
+toolchain go1.24.4
 
 require (
-	github.com/MicahParks/keyfunc/v3 v3.3.10
+	github.com/Masterminds/semver v1.5.0
+	github.com/MicahParks/keyfunc/v3 v3.4.0
 	github.com/alexedwards/argon2id v1.0.0
-	github.com/bufbuild/buf v1.51.0
+	github.com/bufbuild/buf v1.55.1
 	github.com/fsnotify/fsnotify v1.9.0
 	github.com/fzipp/gocyclo v0.6.0
 	github.com/go-critic/go-critic v0.13.0
