@@ -4,6 +4,8 @@ import { By, until, Condition } from 'selenium-webdriver'
 //import * as waitOn from 'wait-on'
 import {
   getRootAndWait,
+  openSidebar,
+  getNavigationLinks,
   takeScreenshotOnFailure,
 } from '../lib/elements.js'
 
@@ -26,17 +28,15 @@ describe('config: empty dashboards are hidden', function () {
     const title = await webdriver.getTitle()
     expect(title).to.be.equal("OliveTin")
 
-    await webdriver.findElement(By.id('sidebar-toggler-button')).click()
+    await openSidebar()
 
-    const navigationLinks = await webdriver.findElements(By.css('#navigation-links a'))
-
-    console.log('navigationLinks', navigationLinks)
-
+    const navigationLinks = await getNavigationLinks()
     expect(navigationLinks).to.not.be.empty
     expect(navigationLinks.length).to.be.equal(1, 'Expected the nav to only have 1 link')
-
+    
     const firstLinkId = await navigationLinks[0].getAttribute('id')
 
     expect(firstLinkId).to.be.equal('showActions', 'Expected the first link to be the actions link')
+
   })
 })
